dashboard/requirements.txt
Flask==3.1.2
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.9
python-dotenv==1.1.1
Werkzeug==3.1.3